# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EcomVina API Documentation",
        "description": "Tài liệu API hệ thống thương mại điện tử EcomVina",
        "contact": {
            "name": "EcomVina Dev Team",
            "email": "fpolydn@fpt.edu.vn"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Server cục bộ"
        },
        {
            "url": "https://ecomvina.yourdomain.com",
            "description": "Server sản phẩm"
        }
    ],
    "paths": {
...
        "/api/danhgias": {
            "get": {
                "tags": [
                    "Đánh giá"
                ],
                "summary": "Lấy danh sách đánh giá",
                "description": "Trả về danh sách các đánh giá với phân trang và tìm kiếm.",
                "operationId": "8bb8fcb931f45a9e411ee27ac327c353",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số lượng phần tử mỗi trang",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Từ khóa tìm kiếm theo nội dung, trạng thái, sản phẩm hoặc người dùng",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách đánh giá trả về thành công"
                    }
                }
            }
        },
        "/api/danhgias/{id}": {
            "get": {
                "tags": [
                    "Đánh giá"
                ],
                "summary": "Xem chi tiết đánh giá",
                "description": "Trả về thông tin chi tiết của một đánh giá theo ID.",
                "operationId": "37603fc16ba8fe9073ffc17e386e1429",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của đánh giá",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chi tiết đánh giá được trả về thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy đánh giá"
                    }
                }
            }
        },
...
    },
...
}